Orders

+++ backend/src/DTOs/Products/Request/ProductRequest.cs
@@ -4,4 +4,4 @@ public sealed record ProductRequest(
     string Name,
     string Category,
     string Image,
-    double Price);
+    decimal Price);
